requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.0.2
pytest==7.4.2